♻️ refactor(fe): createDivForSnowPosition (계산)함수를 컴포넌트에서 utils 폴더로 분리 및 getParticles 함수에서 의존성 주입
